requirements.txt
ultralytics
websockets==13.1
python-telegram-bot 
python-telegram-bot[job-queue]